% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{AB_CIVARTest}
\alias{AB_CIVARTest}
\title{Test on restrictions in the cointegration space}
\usage{
AB_CIVARTest(R0, R1, G, H, h, alpha, beta, Omega, Y0, Z1, Z2)
}
\arguments{
\item{R0}{: T x n matrix containing the residuals of the auxiliary regression of dy_t on Z2_t}

\item{R1}{: T x n matrix containing the residuals of the auxiliary regression of y_(t-1) on Z2_t}

\item{G}{: The restriction matrix on alpha}

\item{H}{: The restriction matrix on beta}

\item{h}{: The restriction vector on beta ( free-varying parameters of beta )}

\item{alpha}{: n x crk matrix containing initial values of the adjustment parameters}

\item{beta}{: n x crk matrix containing initial values of the cointegration vectors}

\item{Omega}{: n x n matrix containing initials of the covariance matrix}

\item{Y0}{: T x n matrix of dY_t}

\item{Z1}{: T x n matrix containing the lagged cointegrated VAR process.}

\item{Z2}{: T x (n(p-1)+#det) matrix containing the lagged stationary regressors of the auxiliary regression}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a list containing (VECMRS, beter, alphar, LSKOEFR, LR, and error)
}\if{html}{\out{</div>}}
}
\description{
This function estimates constrained VECM using the iteration procedure proposed in Boswijk and Doornik (2003)
}
\section{Details}{

This function runs a likelihood ratio test of linear restrictions on $\alpha$ and $\beta$ in a CIVAR model in the following form:
\deqn{vec(\alpha') = G \psi,  vec(\beta) = H\phi + h}

\if{html}{\out{<div class="sourceCode">}}\preformatted{   example 1 (restrictions on alpha) test of exogeneity
             vec(alpha) is 45 x 1  vector  ( n = 9 crk = 5, defined by the model )
             G          is 45 x 40 matrix  ( the first variable is exogenous, i.e. the 5 adjustment coefficient of the first variable are zero )
             psi        40 x 1 vector (free varying parameters not appearing in the specification but implied by G)
             vec(beta)  is 45 x 1 vector   ( N = 0 crk = 5 )
             H          is 45 x 45 identity matrix
             phi        45 x 1 vector (free varying parameters not appearing in the specification but implied by h =0
             h          45 x 1 zero matrix implying ver(beta) = phi  >> no restrictions on beta.
                        (H is identity and h is zero vector implies only restrictions on alpha)

   example 2 (restrictions on beta ) test of PPP
             vec(alpha) is 40 x 1  vector  ( N = 8 crk = 5 ) conditional VECM  or VECMX model
             G          is 40 x 40 identity matrix, implying there is no restriction on alpha
             psi        40 x 1 vector (free varying parameters not appearing in the specification but implied by the identity matrix G
             vec(beta)  is 45 x 1 vector   ( N = 8 crk = 5, defined by the model )
             H          is 45 x 2 matrix that picks out the elements under restrictions ( two columns out of the identity matrix ) a zero row in H and the corresponding h implies zero-restrictions on beta.
                        ones in a row of H and zero in the corresponding h implies non-restricted beta.
             phi        2  x 1  vector (free varying parameters not appearing in the specification but implied by h =0
             h          45 x 1  non zero elements in this vector together with the zero elements in the corresponding row in H are the normalization conditions.
                        (H is identity and h is zero vector implies only restrictions on alpha)
}\if{html}{\out{</div>}}
}

