% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{ABC_MRCIVARestm}
\alias{ABC_MRCIVARestm}
\title{Test of restrictions in the cointegration space of a MRCIVAR model}
\usage{
ABC_MRCIVARestm(res = res, H = H, h = h, phi = phi, G = G, psi = psi)
}
\arguments{
\item{res}{An MRCIVAR object of the output of MRCIGVARDatam}

\item{H}{A matrix specifying restrictions on beta}

\item{h}{A vector specifying restrictions on beta}

\item{phi}{A vector of freely varying parameters in beta}

\item{G}{A matrix specifying restrictions on alpha}

\item{psi}{A vector of freely varying parameters in alpha}
}
\value{
A list containing estimated MRVECM under restrictions, restricted parameter, and likelihood test statistic
}
\description{
Test of restrictions in the cointegration space of a MRCIVAR model
}
