% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{abLRtest2}
\alias{abLRtest2}
\title{Estimation of conditional VECM}
\usage{
abLRtest2(
  y,
  x,
  model = c("I", "II", "III", "IV", "V"),
  bnorm = c("1", "2"),
  type = c("eigen", "trace"),
  p = 1,
  r = 2,
  q = 0.95,
  H = H,
  h = h,
  G = G
)
}
\arguments{
\item{y}{the endogenous variables}

\item{x}{the exogenous variables}

\item{model}{Types of the deterministic specification}

\item{bnorm}{normalization of the cointegration vectors}

\item{type}{types of returns}

\item{p}{Lags of the CIVAR model}

\item{r}{the cointegration rank}

\item{q}{the significance level}

\item{H}{restriction matrix of beta}

\item{h}{restriction vector of beta}

\item{G}{restriction matrix of alpha}
}
\description{
This function runs a test in cointegration space and provides an estimation of restricted alpha and beta.
}
