% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{ABSVAR}
\alias{ABSVAR}
\title{This function solves AB_SVAR from the reduced form and return the A, B matrices and the sum of squared errors}
\usage{
ABSVAR(x0, A0, B0, Sigma)
}
\arguments{
\item{x0}{: difference between the reduced form and the AB form}

\item{A0}{: A matrix in an AB_SVAR model}

\item{B0}{: B matrix in an AB_SVAR model}

\item{Sigma}{: Covariance matrix of the reduced form}
}
\value{
a list contains A B and the difference
}
\description{
This function solves AB_SVAR from the reduced form and return the A, B matrices and the sum of squared errors
}
