% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{ACCIRFconfR}
\alias{ACCIRFconfR}
\title{Calculation of the cumulative impulse response function from a impulse response function}
\usage{
ACCIRFconfR(IRF)
}
\arguments{
\item{IRF}{An impulse response function}
}
\value{
The cumulative impulse response function
}
\description{
Calculation of the cumulative impulse response function from a impulse response function
}
