% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{alpha_perpf}
\alias{alpha_perpf}
\title{This function calculates the orthogonal complementary space}
\usage{
alpha_perpf(alpha = (alpha))
}
\arguments{
\item{alpha}{: a matrix with independent columns}
}
\value{
the orthogonal complementary matrix
}
\description{
This function calculates the orthogonal complementary space
}
