% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_d.R
\name{BoAo2TBoAo}
\alias{BoAo2TBoAo}
\title{A full rank linear transformation of initial Bo and Ao matrices}
\usage{
BoAo2TBoAo(Bo, Ao, W, S)
}
\arguments{
\item{Bo}{The initial Bo matrix}

\item{Ao}{The initial Ao matrix}

\item{W}{A weighting matrix of the GVAR model}

\item{S}{Number of regimes}
}
\value{
Transformed Bo and Ao matrices
}
\description{
This function calculates a full rank linear transformation of initial Bo and Ao matrices with independent common trends and idiosyncratic trends.
}
