% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{BoAoW2G}
\alias{BoAoW2G}
\title{Transformation of country GVAR parameters to the Global VAR parameters}
\usage{
BoAoW2G(Bo, Ao, W, m, n, p)
}
\arguments{
\item{Bo}{parameter matrix of domestic variables}

\item{Ao}{parameter matrix of foreign variables}

\item{W}{the weighting vector of the GVAR model}

\item{m}{the number of variables}

\item{n}{the number of countries}

\item{p}{lags}
}
\value{
An (mn, mn, p ) array of the GVAR(m,n,p) coefficients.
}
\description{
Transformation of country GVAR parameters to the Global VAR parameters
}
