% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{BoAoWs2Gs}
\alias{BoAoWs2Gs}
\title{Calculation of auto regression parameters of a GVAR model from its country equations}
\usage{
BoAoWs2Gs(Bo, Ao, W, m, n, p, state)
}
\arguments{
\item{Bo}{Parmeter matrices of the domestic variables}

\item{Ao}{Parameter matrices of the foreign variables}

\item{W}{The weighting matrix of the GVAR model}

\item{m}{Number of variables of a country}

\item{n}{Number of countries in the GVAR model}

\item{p}{Lag specifications}

\item{state}{A vector specifying state of the GVAR parameters}
}
\value{
The auto regression paramters of the GVAR model
}
\description{
Calculation of auto regression parameters of a GVAR model from its country equations
}
