% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CCIVARData}
\alias{CCIVARData}
\title{Conditional cointegrated VAR}
\usage{
CCIVARData(n1, n2, crk, p, T, type, Bc = NA)
}
\arguments{
\item{n1}{dimension of the conditional cointegrated process}

\item{n2}{dimension of the conditioning variables}

\item{crk}{the cointegration rank}

\item{p}{lag}

\item{T}{number of observations}

\item{type}{types of the deterministic component. "none" and "const" are two options.}

\item{Bc}{Coefficient matrix of the joint cointegration process}
}
\value{
a CCIVAR object which is a list of (n1,n2,p,type,r_np,By,Bx,Cy,Sigma,Y,X,resid,U,check,crk,Bc,Cc)
}
\description{
The function generates data from a conditional cointegrated VAR(p)
}
\examples{

T = 100
res_d <- CCIVARData(n1=4,n2=3,crk=3,p=3,T=T,type="const")
res_e <- CCIVARest(res=res_d)
res_e$Summary


}
