% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CCIVARest}
\alias{CCIVARest}
\title{Estimation of CCIVAR}
\usage{
CCIVARest(res)
}
\arguments{
\item{res}{a CCIVAR object that can be an output of CCIVARData containing at least n1,n2, Y, X, and crk.}
}
\value{
a CCIVAR object with estimated parameter values, AIC, BIC, conditional VECM in a regression format.
}
\description{
This function estimates parameters of a specified conditional cointegrated VAR based on provided data.
}
\examples{
T = 100
res_d <- CCIVARData(n1=4,n2=3,crk=3,p=3,T=T,type="const")
res_e <- CCIVARest(res=res_d)
res_e$Summary

}
