% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CIA2A}
\alias{CIA2A}
\title{Transformation of VECM coefficient matrix to VAR coefficient matrix}
\usage{
CIA2A(CIA)
}
\arguments{
\item{CIA}{a coefficient matrix of VECM}
}
\value{
the coefficient matrix of the corresponding VAR in level
}
\description{
Transformation of VECM coefficient matrix to VAR coefficient matrix
}
