% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_d.R
\name{CIGVAR_Select}
\alias{CIGVAR_Select}
\title{Calculation of information criteria for CIGVAR models}
\usage{
CIGVAR_Select(res, L_V = L_V)
}
\arguments{
\item{res}{: an object generated from CIGVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag and the foreign lag, respectively.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ : a matrix with different lag specifications and the corresponding information criteria.
}\if{html}{\out{</div>}}
}
\description{
This function calculates AIC and BIC criteria for a range of lag specifications of the domestic and foreign variables up to the maximum given in L_V.
}
\examples{


n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
p[2,1] = 2;  p[3,1] = 2
res_d = CIGVARData(m=2,n=5,p=p,T=5000,type="const",DFYflag=0)
res_d$r_npo
res_d$check

plot(ts(res_d$Y[1:100,1:10]))

res_e = CIGVARest(res_d)
res_e$Summary
res_e$crk = c(1,1,1,1,1)
L_V = c(4,4)
CIGVARSelecte = CIGVAR_Select(res=res_e,L_V=c(4,4))
CIGVARSelecte[which.min(CIGVARSelecte[,17]),]
CIGVARSelecte[which.min(CIGVARSelecte[,19]),]

}
