% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_c.R
\name{CIGVARData1}
\alias{CIGVARData1}
\title{Data generating process CIGVAR(n,m,p)}
\usage{
CIGVARData1(
  m,
  n,
  p,
  T,
  W = NA,
  r_npo = NA,
  Ao = NA,
  Bo = NA,
  Co = NA,
  Uo = NA,
  Sigmao = NA,
  type = NA,
  X = NA,
  mu = NA,
  d = 0,
  crk = NA,
  Ncommtrend = 1
)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: an (n x 3) matrix, each raw contains the lag length of the domestic variables, the lag length foreign variables, and the number of the exogenous variables.}

\item{T}{: number of observations.\preformatted{           (m,n,p,T) are parameters which must be provided.
}}

\item{W}{: n x n weighting matrix. w_ij is the weight of foreign country j in the foreign variables of i-th country diag(W)=0}

\item{r_npo}{: m x p x n array collecting the roots of the characteristic functions of the country VAR. The number of ones i-th row of r_npo is the number of unit roots i-th country/unit.}

\item{Ao}{: m x m x p x n array collecting the off-diagonal block of coefficients which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: m x m x p x n array collecting the n country VAR(p) coefficients.  Bo are coefficients of stationary domestic VAR(p).}

\item{Co}{: m x (k+1) x n array collecting the coefficients of the deterministic components of the n countries.}

\item{Uo}{: an T x mn matrix of the temporally independent innovation processes}

\item{Sigmao}{: mn x mn matrix of the covariance matrix of the GVAR(m,n,p)\preformatted{             (W,r_npo,Ao,Bo,Uo,Sigmao) if not provided, they will be generated randomly.
}}

\item{type}{: deterministic component "const" and "none" are two options}

\item{X}{: (T x k) matrix of exogenous variables.}

\item{mu}{: if type = "const" an mn vector of drifts of the CIGVAR(p) process}

\item{d}{: d = 0 indicates foreign variables do not enter the cointegration space. d = 1 allows foreign variables to enter the cointegration space.}

\item{crk}{: an n-vector containing the cointegration rank in each country/unit. crk is used in estimation.}

\item{Ncommtrend}{: number of common stochastic trends across all countries. Ncommtrend=0 for the case of no common trend.}
}
\value{
\preformatted{ a CIGVAR object containing the generated data, the used parameters and the exogenous variables.
}
}
\description{
This function generates data from a cointegrated global VAR process with common and idiosyncratic stochastic trends.
}
\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
r_npo = (1:(3 * 3 * 5))*0; dim(r_npo) = c(3,3,5)
r_npo[,,1] = matrix(c(1,1,3,2,3,3,3,3,3),3,3)
r_npo[,,2] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,3] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,4] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,5] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)

res_d = CIGVARData1(m=3,n=5,p=p,T=500,r_npo=r_npo,type="const",d=0,Ncommtrend=1)
max(res_d$Y)
STAT(res_d$G)
plot(ts(res_d$Y[,1:10]))
res_e = CIGVARest(res_d)
res_e$Summary
## adjust the model specification according to the cointegration test results
res_d$crk = c(1,2,2,2,2)
res_e = CIGVARest(res_d)
res_e$Summary
#### unit roots in the CIGVAR model
STAT(res_e$G)
}
