% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_c.R
\name{CIGVARest}
\alias{CIGVARest}
\title{Estimation of CIGVAR(m,n,p)}
\usage{
CIGVARest(res)
}
\arguments{
\item{res}{a CIGVAR object of an output of CIGVARData or CIGVARData1 including at least values of m,n,p,type,Y,crk and optionally X.}
}
\value{
a CIGVAR object with estimated parameter values, AIC, BIC and LH
}
\description{
This function estimates the parameters of a CIGVAR(m,n,p) object based on provided data.
It runs a VECM estimation country/unit by country/unit under a given cointegration rank and pieces the results together to obtain a CIGVAR.
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
r_npo = (1:(3 * 3 * 5))*0; dim(r_npo) = c(3,3,5)
r_npo[,,1] = matrix(c(1,1,3,2,3,3,3,3,3),3,3)
r_npo[,,2] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,3] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,4] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,5] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)

res_d = CIGVARData1(m=3,n=5,p=p,T=500,r_npo=r_npo,type="const",d=0,Ncommtrend=1)
max(res_d$Y)
STAT(res_d$G)
plot(ts(res_d$Y[,1:10]))
res_e = CIGVARest(res_d)
res_e$Summary
res_e$Summary$CRK_Test
## adjust the model specification according to the cointegration rank test
res_d$crk = c(1,2,2,2,2)
res_e = CIGVARest(res_d)
res_e$Summary
## unit roots in the CIGVAR model
STAT(res_e$G)
}
