% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CIVARData}
\alias{CIVARData}
\title{Data generating process of CIVAR(p)}
\usage{
CIVARData(n, p, T, r_np, A, B, Co, C1, U, Sigma, type, X, mu, Yo, crk)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: lag length}

\item{T}{: number of observations}

\item{r_np}{: an (n x p) matrix of roots of the characteristic polynomials of n independent AR(p)-processes. An element 1 in a row implies a unit root process. If the matrix is not provided, it will be generated randomly with one unit root in the first row.}

\item{A}{: an (n x n) full rank transformation matrix that is used to generate correlated a CIVAR(p) from the n independent AR(p) processes with unit roots.}

\item{B}{: an (n,n,p) array containing the coefficients of the CIVAR(p) process. If B is not given, it will be calculated out of r_np and A.}

\item{Co}{: an (n,k+1) matrix containing coefficients of deterministic components in a CIVAR(p) process. For type="none" Co = 0*(1:n), for type="const" Co is an n vector, for type="exog0" Co is a (n,k) matrix, and for type="exog1" Co is an (n,1+k) matrix.}

\item{C1}{: an (n,1) matrix containing coefficients of the trend component. C1 = 0 if there is no trend in the data.}

\item{U}{: residuals, if it is not NA it will be used as input to generate the CIVAR(p)}

\item{Sigma}{: an n x n covariance matrix of the CIVAR(p)}

\item{type}{: types of deterministic components. "none","rconst","const","rtrend", "trend", "exog0" and "exog1" are 7 options
\itemize{
\item{none: }{No deterministic component in VECM (Case I)}
\item{rconst: }{An intercept restricted in the cointegration space (Case II)}
\item{const: }{An unrestricted intercept (Case III)}
\item{rtrend: }{An trend component restricted in the cointegration space and an unrestricted intercept in VECM (Case IV)}
\item{trend: }{An unrestricted trend component and an unretricted intercept in VECM (Case V)}
\item{exog0: }{Model with exogeneous variable and no deterministic components}
\item{exog1: }{Model with exogeneous variable and an unrestricted intercept}
}}

\item{X}{: a (T x k) matrix of exogenous variables.}

\item{mu}{: an n vector of drifts of the CIVAR(p) process}

\item{Yo}{: (p x n) matrix of initial values of the process}

\item{crk}{: number of cointegration relations. It equals n-r, where r is the number of unit roots.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ An object of CIVAR(p) containing the generated data, the parameters used and the exogenous variables. res = list(n,p,type,r_np,Phi,A,B,Co,Sigma,Y,X,resid,U,Y1,Yo,check)
}\if{html}{\out{</div>}}
}
\description{
This function generates data from a cointegrated vector autoregressive process CIVAR(p) and returns CIVAR(p) object that is a list containing data and parameters used in the CIVAR(p) process.
}
\examples{
res_d = CIVARData(n=2,p=2,T=100,type="const")
res_d = CIVARData(n=2,p=2,T=10,Co=c(1:2)*0,type="none")
res_d = CIVARData(n=2,p=2,T=10,Co=c(1:2)*NA,  type="const")

p = 3
n = 4
r_np = matrix(c(1,2,1.5,1.5,2.5,2.5,2,-1.5,2,-4,1.9,-2.1),4,3)

res_d  = CIVARData(n=4,p=3,T=200,r_np=r_np,Co=matrix(0,n,1) ,type="none",crk=3)
res_e  = CIVARest(res=res_d)
sum(abs(res_e$B-res_d$B))
sum(abs(res_e$Co-res_d$Co))


res_d  = CIVARData(n=4,p=3,T=200)
res_e = CIVARest(res=res_d)
sum(abs(res_e$B-res_d$B))
sum(abs(res_e$Co-res_d$Co))
plot(ts(res_d$Y))
}
