% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CIVARest}
\alias{CIVARest}
\title{Estimation of CIVAR(p)}
\usage{
CIVARest(res)
}
\arguments{
\item{res}{:an object of CIVAR(p) containing the components which are the output of CIVARData including at least: n, p, Y, crk, and optionally X and type.}
}
\value{
res  an object of CIVAR(p) containing estimated parameter values, AIC, BIC, LH and the estimated VECM in regression format.
}
\description{
This function estimates parameters of a specified CIVAR(p) model based on provided data.
}
\examples{
p = 3
n = 4
r_np = matrix(c(1,2,1.5,1.5,2.5,2.5,2,-1.5,2,-4,1.9,-2.1),4,3)

res_d  = CIVARData(n=4,p=3,T=200,r_np=r_np,Co =(1:n)/(1:n)*0,type="none",crk=3)
res_e  = CIVARest(res=res_d)
res_d  = CIVARData(n=4,p=3,T=200)
B = res_d$B
plot(ts(res_d$Y))
res_d$Co
res_d$type
res_d$crk
res_e = CIVARest(res=res_d)



}
