% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_c.r
\name{CIVARTest}
\alias{CIVARTest}
\title{Test of restrictions in the cointegration space and estimation of restricted VECM}
\usage{
CIVARTest(res = res, H = H, h = h, phi = phi, G = G, Dxflag = 0)
}
\arguments{
\item{res}{an estimated CIVAR object with without restrictions}

\item{H}{the restriction matrix on beta}

\item{h}{the restriction vector on beta ( free-varying parameters of beta )}

\item{phi}{the restriction vector on beta}

\item{G}{the restriction matrix on alpha}

\item{Dxflag}{A flag that indicates whether X enters the cointegration space.}
}
\value{
a list containing (VECMRS, beter, alphae, LSKOEFR, LR and error)
}
\description{
This function estimates constrained VECM using the iteration procedure proposed in Boswijk and Doornik (2003)
}
\section{Details}{

This function runs a likelihood ratio test of linear restrictions on alpha and beta in a CIVAR model in the following form:
\deqn{vec(\alpha') = G \psi,  vec(\beta) = H\phi + h}\preformatted{   example 1 (restrictions on alpha) test of exogeneity
             vec(alpha) is 45 x 1  vector  ( n = 9 crk = 5, defined by the model )
             G          is 45 x 40 matrix  ( the first variable is exogenous, i.e. the 5 adjustment coefficient of the first variable are zero )
             psi        40 x 1 vector (free varying parameters not appearing in the specification but implied by G)
             vec(beta)  is 45 x 1 vector   ( N = 0 crk = 5 )
             H          is 45 x 45 identity matrix
             phi        45 x 1 vector (free varying parameters not appearing in the specification but implied by h =0
             h          45 x 1 zero matrix implying ver(beta) = phi  >> no restrictions on beta.
                        (H is identity and h is zero vector implies only restrictions on alpha)

   example 2 (restrictions on beta ) test of PPP
             vec(alpha) is 40 x 1  vector  ( N = 8 crk = 5 ) conditional VECM  or VECMX model
             G          is 40 x 40 identity matrix, implying there is no restriction on alpha
             psi        40 x 1 vector (free varying parameters not appearing in the specification but implied by the identity matrix G
             vec(beta)  is 45 x 1 vector   ( N = 8 crk = 5, defined by the model )
             H          is 45 x 2 matrix that picks out the elements under restrictions ( two columns out of the identity matrix ) a zero row in H and the corresponding h implies zero-restrictions on beta.
                        ones in a row of H and zero in the corresponding h implies non-restricted beta.
             phi        2  x 1  vector (free varying parameters not appearing in the specification but implied by h =0
             h          45 x 1  non zero elements in this vector together with the zero elements in the corresponding row in H are the normalization conditions.
                        (H is identity and h is zero vector implies only restrictions on alpha)
}
}

\examples{
X = matrix(stats::rnorm(2*207),207,2)
colnames(X) = c("ex1","ex2")
res_d  = CIVARData(n=6,p=2,T=207,crk=2,type="exog1", X=as.matrix(X))
colnames(res_d$Y) =  c("w","p","U_l","r","yn","y")
res_e =  CIVARest(res=res_d)
res_e$tst$erg
res_e$tst$beta
res_e$tst$VECMS
n = 6;  crk = 2
G = diag(n*crk); G0 = G; psi = matrix(1,n*crk,1); psi0=psi;   ### No restrictions on alpha
H = diag(n*crk); H2 = H[,-seq(1,n*crk,n)];
h2 = matrix(0,n*crk,1);
h2[seq(1,n*crk,n),1] = rep(1,crk);
phi2 = matrix(1,(n-1)*crk,1)                               ### no restrictions but normalization

G0\%*\%psi0; H2\%*\%phi2+h2

ABtest = CIVARTest(res=res_d,H=H2,h=h2,phi=phi2,G=G0,Dxflag=0)
ABtest$betar
ABtest$alphar
ABtest$VECMR$coefficients
ABtest$LR
#1-pchisq(ABtest$LR,2)   ### the fourth is the used restrictions


res_d  = CIVARData(n=7,p=2,T=207,crk=2,type="const")
colnames(res_d$Y) =  c("w","p","U_l","r","yn","y","fsi")
res_e =  CIVARest(res=res_d)
res_e$tst$erg
res_e$tst$beta
res_e$tst$VECMS

n = 7;  crk = 2
G = diag(n*crk); G0 = G; psi = matrix(1,n*crk,1); psi0=psi;   ### No restrictions on alpha

H = diag(n*crk); H2 = H[,-c(1,8)]; h2 = matrix(0,n*crk,1);
h2[c(1,8),1] = c(1,1);  phi2 = matrix(1,12,1)                 ### normalization

G0\%*\%psi0; H2\%*\%phi2+h2

ABtest = CIVARTest(res=res_d,H=H2,h=h2,phi=phi2,G=G0,Dxflag=1)
ABtest$LR
ABtest$betar
ABtest$alphar
#1-pchisq(ABtest$LR,4)


}
