% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{cvecm}
\alias{cvecm}
\title{Estimation of a conditional vector error correction process.}
\usage{
cvecm(
  y,
  x,
  model = c("I", "II", "III", "IV", "V"),
  type = c("eigen", "trace"),
  crk = 2,
  p = 1,
  q = 0.95
)
}
\arguments{
\item{y}{data matrix of the endogenous variables}

\item{x}{data matrix of the exogenous variables}

\item{model}{type of the specification of the deterministic components. "none" and "const" are two options.}

\item{type}{=c("eigen", "trace")}

\item{crk}{cointegration rank}

\item{p}{lag length in level}

\item{q}{significance level}
}
\value{
a list of estimated parameters and test statistics
}
\description{
This function estimates the unknown parameters of a multi regime conditional VECM based on provided data.
}
