% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{CVECM2CVAR}
\alias{CVECM2CVAR}
\title{Transformation of the estimated parameters of a conditional VECM into the parameters of the corresponding conditional VAR}
\usage{
CVECM2CVAR(param, beta, p = c(1, 2, 2, 2, 2), s = NA, N2)
}
\arguments{
\item{param}{estimated parameters of a conditional VECM}

\item{beta}{the estimated cointegration vecters}

\item{p}{a vector specifying different types of conditional VECM}

\item{s}{indicator variable of different regimes}

\item{N2}{dimension of the conditioning variables}
}
\value{
A list of (A, B, C) with B the conditional VAR parameter matrix, C the parameter of the deterministic components, and A the parameter matrix of foreign variables for CIGVAR.
}
\description{
Transformation of the estimated parameters of a conditional VECM into the parameters of the corresponding conditional VAR
}
