% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{Embed}
\alias{Embed}
\title{Embedding a time series}
\usage{
Embed(y = tseries::ts(c(1:20)), p = 3, prefix = "")
}
\arguments{
\item{y}{The time series to be embedded}

\item{p}{Number of lags for embedding}

\item{prefix}{Prefix of the column names}
}
\value{
A matrix containing the embedded time series y.
}
\description{
Embeds the time series y into a low-dimensional Euclidean space with column names.
}
