% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{f}
\alias{f}
\title{An auxiliary function for nonlinear estimation of the cointegration vectors}
\usage{
f(x, beta, Z1, St, NSt, Y0, Z2)
}
\arguments{
\item{x}{A vector containing the parameters in the cointegration vectors.}

\item{beta}{The initial value of the cointegration vectors}

\item{Z1}{The lagged I(1) series of the model}

\item{St}{The indicator series of regime 1}

\item{NSt}{The indicator series of regime 2}

\item{Y0}{The first difference series}

\item{Z2}{The lagged difference series}
}
\description{
An auxiliary function for nonlinear estimation of the cointegration vectors
}
