% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{f_constrained}
\alias{f_constrained}
\title{Help function for nonlinear optimization in constrained estimation of CIVAR}
\usage{
f_constrained(
  x,
  beta = beta,
  alpha = alpha,
  G = G,
  H = H,
  phi = phi,
  psi = psi,
  h = h,
  Z1,
  St,
  NSt,
  Y0,
  Z2
)
}
\arguments{
\item{x}{Optimization variables}

\item{beta}{Cointegration vectors}

\item{alpha}{Adjustment vectors}

\item{G}{A matrix specifying restrictions on alpha}

\item{H}{A matrix specifying restrictions on beta}

\item{phi}{Freely varying parameters in beta}

\item{psi}{Freely varying parameters in alpha}

\item{h}{A vector specifying restrictions in beta}

\item{Z1}{I(1) data matrix}

\item{St}{Regime 1 indicator series}

\item{NSt}{Regime 2 indicator series}

\item{Y0}{I(0) data matrix}

\item{Z2}{I(0) data matrix}
}
\value{
Sum of squared residuals
}
\description{
Help function for nonlinear optimization in constrained estimation of CIVAR
}
