% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FSIIO}
\alias{FSIIO}
\title{Financial stress index and industrial output}
\format{
An object of class \code{data.frame} with 370 rows and 30 columns.
}
\usage{
FSIIO
}
\description{
Financial stress index and industrial output
}
\keyword{datasets}
