% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{girf_MRCIVAR_RM_CB}
\alias{girf_MRCIVAR_RM_CB}
\title{Generalized impulse response functions of MRCIVAR(n,p,S) with regime migrations with a confidence interval}
\usage{
girf_MRCIVAR_RM_CB(
  res,
  shock,
  R,
  nstep,
  Omega_hist = NA,
  resid_method = "parametric",
  conf_level,
  N
)
}
\arguments{
\item{res}{: an MRCIVAR object containing the components  of an output of MRCIVARestm1.}

\item{shock}{: an n vector containing the shocks as impulse.}

\item{R}{: the number runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: the length of the responses}

\item{Omega_hist}{: a (P x n) matrix of initial values, from which the impulse response functions start. Omega_hist determines from which regime the impulse response functions start. For Omega_hist=NA, the impulse response functions will start from the most resent observations.}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric bootstrap.}

\item{conf_level}{: a vecter contain the level of confidences}

\item{N}{: the number of bootstrap runs in containing the bootstrapped confidence intervals.}
}
\value{
an (n, n, nstep+1,3) array containing the impulse response functions with lower and upper confidence bonds. The rows represent response, and the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRVAR(n,p,S) with a bootstrapping confidence interval.
}
\details{
For a given shock vector SHCK:

GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))

See H.H. Pesaran and Y. Shin (1998) Generalized impulse response analysis in linear multivariate models, Economics Letters, 58(1) p. 17-29.
and G. Koop, M. H. Pesaran, and S. M. Potter (1996), Impulse response analysis in nonlinear multivariate models, Journal of Econometrics, 74 (1996) 119-74.
###########################################################################################################################################
}
\examples{
n =4

Sigma = 1:(n*n*2)
dim(Sigma) = c(n,n,2)
Sigma[,,1] = diag(n)
Sigma[,,2] = diag(n)
p=matrix(0,2,2)
p[,1] = c(3,3)
res_d = MRCIVARDatam(n=n,p=p,T=250,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=2)
#colnames(res_d$Y) = c("w","p","I","Q")
max(abs(res_d$Y))
res_e = MRCIVARestm1(res=res_d)
res_e$Summary
Mod(STAT(res_e$B[,,,1]))
Mod(STAT(res_e$B[,,,2]))

#res_e$Summary
if (! max(Mod(STAT(res_e$Bo[,,,1])),Mod(STAT(res_e$Bo[,,,2])) ) > 1.000000001 ) {

 IRF  = irf_MRCIVAR_CB(res_e,nstep=20,irf="gen1",runs=10,comb=NA,G=NA,conf=c(0.05,0.95))
 IRF_g1 <- IRF_graph(IRF[[1]])        # irf of regime 1
 IRF_g2<- IRF_graph(IRF[[2]])         # irf of regime 2

}

}
