% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{girf_MRVAR_RM}
\alias{girf_MRVAR_RM}
\title{Generalized impulse response functions of MRVAR(n,p,S) with regime migrations}
\usage{
girf_MRVAR_RM(RES, shock, R, nstep, Omega_hist = NA, resid_method)
}
\arguments{
\item{RES}{: an MRVAR object that is an output of MRVARest.}

\item{shock}{: an n vector containing the shocks as impulse.}

\item{R}{: the number runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: the length of the responses}

\item{Omega_hist}{: a (P x n) matrix of initial values, from which the impulse response functions start. Omega_hist determines from which regime the impulse response functions start. For Omega_hist=NA, the impulse response functions will start from the most resent observations.}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric bootstrap.}
}
\value{
an (n x n x nstep+1) matrix of impulse response functions. The rows represent response the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRVAR(n,p,S) for a given shock vector SHCK.
}
\details{
\subsection{GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))}{

See H.H. Pesaran and Y. Shin (1998) Generalized impulse response analysis in linear multivariate models, Economics Letters, 58(1) p. 17-29.
and G. Koop, M. H. Pesaran, and S. M. Potter (1996), Impulse response analysis in nonlinear multivariate models, Journal of Econometrics, 74 (1996) 119-74.
###########################################################################################################################################
}
}
\examples{
p = matrix(c(2,1,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=300,S=2,SESVI=1,type="none")
max(res_d$Y)
colnames(res_d$Y) = c("R","P")
res_e = MRVARest(res=res_d)
RF3 = girf_MRVAR_RM(RES=res_e,shock=c(1,1),R = 100,nstep=20,Omega_hist=NA,resid_method="parametric")
RF4 = girf_MRVAR_RM_CB(RES=res_e, shock=c(1,1), R=100, nstep=20, Omega_hist=NA,
resid_method = "parametric", conf_level=c(0.05,0.95), N=100)
IRF_list <-IRF_graph(RF4)

}
