% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{girf_MRVAR_RM_CB}
\alias{girf_MRVAR_RM_CB}
\title{Generalized impulse response functions of MRVAR(n,p,S) with regime migrations and confidence bands}
\usage{
girf_MRVAR_RM_CB(
  RES,
  shock,
  R,
  nstep,
  Omega_hist = NA,
  resid_method = "parametric",
  conf_level,
  N
)
}
\arguments{
\item{RES}{: an estimated MRVAR object.}

\item{shock}{: an n-vector containing the shocks as impulse.}

\item{R}{: number of runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: length of the impuse response function}

\item{Omega_hist}{: a (P x n) matrix of initial values, from which the impulse response functions start. Omega_hist determines from which regime the impulse response functions start.}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric bootstrap.}

\item{conf_level}{: a vector contain the level of confidences}

\item{N}{: the number of bootstrap runs in containing the bootstrapped confidence intervals.}
}
\value{
an (n, n, nstep+1,3) array containing the impulse response functions with lower and upper confidence bonds. The rows represent response, and the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRVAR(n,p,S) for given a shock vector. It provides also a bootstrapped confidence interval.
}
\details{
\subsection{GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))}{

See H.H. Pesaran and Y. Shin (1998) Generalized impulse response analysis in linear multivariate models, Economics Letters, 58(1) p. 17-29.
and G. Koop, M. H. Pesaran, and S. M. Potter (1996), Impulse response analysis in nonlinear multivariate models, Journal of Econometrics, 74 (1996) 119-74.
}
}
\examples{
p = matrix(c(2,1,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=300,S=2,SESVI=1,type="none")
max(res_d$Y)
colnames(res_d$Y) = c("R","P")
res_e = MRVARest(res=res_d)
RF3 = girf_MRVAR_RM(RES=res_e,shock=c(1,1),R = 200,nstep=20,Omega_hist=NA,resid_method="parametric")
RF4 = girf_MRVAR_RM_CB(RES=res_e, shock=c(1,1), R=200, nstep=20, Omega_hist=NA,
resid_method = "parametric", conf_level=c(0.05,0.95), N=100)
IRF_list <-IRF_graph(RF4)
}
