% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{GVAR_Select}
\alias{GVAR_Select}
\title{GVAR lag selection}
\usage{
GVAR_Select(res = res, L_V = L_V, I = I)
}
\arguments{
\item{res}{: a GVAR object obtained from GVARData or GVARest.}

\item{L_V}{: a two components vector containing the maxima of the domestic lag and the foreign lag, respectively.}

\item{I}{: Index of the country under investigation.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ : A matrix with different lag specifications and values of the model information criteria.
}\if{html}{\out{</div>}}
}
\description{
Calculation of the information criteria of a GVAR country models for a given range of maximum lags.
}
\examples{

n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1; p[2:3,2] = 2
res_d = GVARData(m=2,n=4,p=p,T=4000,type="const")


I = 3
L_V = c(4,4)
res_d$p
GVARSelect = GVAR_Select(res=res_d,L_V=c(4,4),I=2)
GVARSelect[which.min(GVARSelect[,3]),]
GVARSelect[which.min(GVARSelect[,4]),]

}
