% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{GVARData}
\alias{GVARData}
\title{Data generating process of GVAR(m,n,p)}
\usage{
GVARData(
  m,
  n,
  p,
  T,
  W = NA,
  r_npo = NA,
  Ao = NA,
  Bo = NA,
  Co = NA,
  Uo = NA,
  Sigmao = NA,
  type = NA,
  X = NA,
  mu = NA
)
}
\arguments{
\item{m}{: number of variables}

\item{n}{: number of countries/units}

\item{p}{: an (n x 3) matrix in which each row contains the lag length of the domestic variables, the lag length of the foreign variables, and the number of exogenous variables.}

\item{T}{: number of observations

\if{html}{\out{<div class="sourceCode">}}\preformatted{           (m,n,p,T) are parameters which must be provided.
}\if{html}{\out{</div>}}}

\item{W}{: an (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of i-th country diag(W)=0}

\item{r_npo}{: an (m, p, n) array collecting the roots of the characteristic polynomials in Lags for each of the m dimensional  variable across n countries.}

\item{Ao}{: an (m, m, p, n) array collecting the off-diagonal block of coefficients that are coefficients of the foreign variables.}

\item{Bo}{: an (m, m, p, n) array collecting the coefficients of the domestic variables.}

\item{Co}{: an (m, k+1, n) array collecting the coefficients of the deterministic components of the n countries.}

\item{Uo}{: an (T x mn) matrix of the temporally independent innovations}

\item{Sigmao}{: an (mn x mn) matrix of the covariance matrix of the GVAR(m,n,p)

\if{html}{\out{<div class="sourceCode">}}\preformatted{           (W,r_npo,Ao,Bo,Uo,Sigmao) if not provided, they will be generated randomly.
}\if{html}{\out{</div>}}}

\item{type}{: types of deterministic components: "const", "none", "exog0", and "exog1" are the options.}

\item{X}{: (T x k x n ) array of exogenous variables.}

\item{mu}{: if type = "const" mu has the same dimension as Co. It contains the means of the time series in the system.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a GVAR object which is a list("Y","X","Uo","G","C","Sigmao","r_npo","Ao","Bo","Co","W","m","n","p","mu","check","type") containing the generated data, the used parameters, and the inputted of exogenous variables.
}\if{html}{\out{</div>}}
}
\description{
This function generates data from a stationary global vector auto regressive process and return a GVAR(m,n,p) object containing data and parameters used in the GVAR(m,n,p) process.
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 2; p[,2]=1;
res_d = GVARData(m=2,n=5,p=p,T=100,type="const")
max(res_d$Y)
dim(res_d$Y)
res_e = GVARest(res = res_d)
res_e$Summary

X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X3 = matrix(rnorm(200),200,1)
X4 = matrix(rnorm(200),200,1)
X  = cbind(X1,X2,X3,X4)
dim(X) = c(200,1,4)
n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1;   p[,3]=1; p[2,2]=2;
p    ## country-wise lag specification

res_d = GVARData(m=2,n=4,p=p,T=200,type="exog0",X=X)
res_e = GVARest(res = res_d)
res_e$Summary

IRF_CB = irf_GVAR_CB(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
dim(IRF_CB)
IRF_g = IRF_graph(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)

}
