% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{GW2BoAo}
\alias{GW2BoAo}
\title{Transformation of a global VAR parameter matrix to the country GVAR parameter matrices}
\usage{
GW2BoAo(G, W)
}
\arguments{
\item{G}{The GVAR parameter matrix}

\item{W}{The weighting matrix of the GVAR model}
}
\value{
A list containing the parameter matrices of domestric variables and foreign variables
}
\description{
Transformation of a global VAR parameter matrix to the country GVAR parameter matrices
}
