% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{INVI}
\alias{INVI}
\title{Inverse of a partial covariance matrix}
\usage{
INVI(sigma, c, i)
}
\arguments{
\item{sigma}{The input covariance matrix}

\item{c}{The weighting vector of a concerted policy action}

\item{i}{The index of responding variables}
}
\value{
A response matrix
}
\description{
Inverse of a partial covariance matrix
}
