% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{irf_B_sigma}
\alias{irf_B_sigma}
\title{Impulse response function of a vector autoregressive model}
\usage{
irf_B_sigma(
  B,
  sigma,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "genN1", "comb1", "smat", "concerts1",
    "PTdecomp", "ABSVAR", "irfX"),
  G = NA,
  A0 = NA,
  B0 = NA,
  smat = NA,
  Xshk = NA
)
}
\arguments{
\item{B}{An (nxnxp) coefficients array of an n-dimensional VAR(p) model}

\item{sigma}{The covariance matrix of the VAR(p) model}

\item{nstep}{Number of steps of the impulse response function}

\item{comb}{An n-vector of weights of the coordinated policy actions}

\item{irf}{Type of impulse response function}

\item{G}{The matrix used in the permanent and transitory decomposition}

\item{A0}{The matrix for A/B identification in VAR}

\item{B0}{The matrix for A/B identification in VAR}

\item{smat}{An explicit decomposition matrix that defines a structural shock.}

\item{Xshk}{The shock matrix of a one unit exogenous shock}
}
\description{
This function generates impulse response functions for VAR,CIVAR,MRVAR MRCIVAR, also for GVAR, CIGVAR, MRGVAR and MRCIGVAR.
For the later four classes of models it also provides the functionalities to calculate the global, regional and country-specific shocks.
It also calculates global and regional responses and coordinated policy actions.
}
