% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_d.R
\name{irf_CIGVAR_CB}
\alias{irf_CIGVAR_CB}
\title{Impulse Response Functions of CIGVAR(m,n,p)}
\usage{
irf_CIGVAR_CB(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res}{: a CIGVAR object of an output of CIGVARest.}

\item{nstep}{: length of the impulse response functions}

\item{comb}{: an mn-vector specifying combined impulse such as global shocks, regional shocks, or concerted actions.}

\item{irf}{: types of the impulse response functions. irf=c("gen","chol","chol1","gen1","comb1"), gen for generalized IRF with one standard deviation shocks, gen1 for generalized IRF with one unit impulse, chol for IRF with Cholezky decomposition of the covariance matrix, chol1 for Cholezky decomposition with one unit impulse, comb1 for concerted action with one unit impulse.}

\item{runs}{: number of bootstrap runs to generate the confidence bands}

\item{conf}{: a two component vector containing the tail probabilities of the bootstrap confidence interval}
}
\value{
a matrix of (mn,mn,nstep,3) as the IRF columns representing the impulse rows the responses
}
\description{
This function generates impulse response functions of an estimated CIGVAR(m,n,p) with confidence bands.
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
p[2,1] = 2;  p[3,1] = 2
res_d = CIGVARData(m=2,n=5,p=p,T=500,type="const",DFYflag=0)
res_d$r_npo
res_d$check
res_e = CIGVARest(res_d)
res_e$Summary

### Impulse response function

IRF_CB = irf_CIGVAR_CB(res=res_e,nstep=20,comb=NA,irf="gen1",runs=200,conf=c(0.05,0.95))

dim(IRF_CB)
IRF_g = IRF_graph(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)




}
