% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_c.r
\name{irf_CIVAR_CB}
\alias{irf_CIVAR_CB}
\title{Impulse response of an estimated CIVAR.}
\usage{
irf_CIVAR_CB(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1", "PTdecomp", "ABSVAR", "irfX"),
  G = NA,
  A0 = NA,
  B0 = NA,
  Xshks = NA,
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res}{: a CIVAR object such as an output of CIVARest.}

\item{nstep}{: the length of the impulse response functions.}

\item{comb}{: a weighting matrix specifying the weights used in the impulse response functions of a global VAR. Its default value is NA for CIVAR(p).}

\item{irf}{: types of the generated impulse response functions.}

\item{G}{: the transformation matrix for PTdecomp}

\item{A0}{: the transformation matrix for AB identification}

\item{B0}{: the transformation matrix for AB identification}

\item{Xshks}{: the number of selected exogenous shocks}

\item{runs}{: number of runs used in the the calculation of the bootstrap confidence interval.}

\item{conf}{: a two component vector containing the tail probabilities of the bootstrap confidence interval.}
}
\value{
\preformatted{  an array of dimension (n, n, nstep, 3).
}
}
\description{
This function generates the impulse response functions of an estimated CIVAR model
}
\examples{

res_d = CIVARData(n=4,p=2,T=84,Co=matrix(c(1,1,1,1),4,1)*0,type="none",crk=1)
res_e = CIVARest(res=res_d)
res_e$Summary

IRF_CB = irf_CIVAR_CB(res=res_e, nstep=20, comb=NA, irf = "gen1", runs = 20, conf = c(0.05, 0.95))
IRF_g = IRF_graph(IRF_CB)

IRF_CB     = irf_CIVAR_CB(res=res_e, nstep=30, comb=NA, irf = "PTdecomp", G = NA, A0=NA,B0=NA,
 runs = 20, conf = c(0.05, 0.95))
IRF_g = IRF_graph(IRF_CB)
# The first three shocks have permanent effects,
# while the fourth shock does not have a permanent effect.



T=100
X=matrix(rnorm(2*T),T,2)
res_d = VARData(n=3,p=2,T=T,Co=matrix(c(0,0,0,1,2,3,3,2,1),3,3), type="exog0",X=X) ;
res_e = VARest(res=res_d);
res_d$Co
res_e$Summary
IRF_CB = irf_VAR_CB(res=res_e,nstep=20, comb=NA, irf = "irfX", Xshks=c(1:2),
runs = 100, conf = c(0.05, 0.95))
IRF_list <-IRF_graph(IRF_CB,Names =c("Y1","Y2","Y3"),INames=c("X1","X2"),
response = c(1:3), impulse = c(1:3), n = 3)

}
