% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{irf_GloabalResponse_CB}
\alias{irf_GloabalResponse_CB}
\title{Global Responses}
\usage{
irf_GloabalResponse_CB(IRF_CB, comb_all)
}
\arguments{
\item{IRF_CB}{an output of irf_MRGVAR_CB}

\item{comb_all}{a weighting matrix for the aggregation of the global responses}
}
\value{
a list containing the global impulse response functions and the accumulated global impulse response functions.
}
\description{
This function calculates global or regional responses from a set of bootstrapped impulse response functions and a weighting matrix for the aggregation of the global responses.
}
