% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{irf_GVAR_CB}
\alias{irf_GVAR_CB}
\title{Impulse Response Functions of GVAR}
\usage{
irf_GVAR_CB(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res}{: an object of GVAR that is a list of the output of GVARest}

\item{nstep}{: length of the impulse response functions}

\item{comb}{: an mn vector specifying combined impulse such as global shocks, regional shocks, or concerted actions.}

\item{irf}{: types of the impulse response functions. irf=c("gen","chol","chol1","gen1","comb1"), gen for generalized IRF with one standard deviation shocks, gen1 for generalized IRF with one unit impulse, chol for IRF with Cholezky decomposition of the covariance matrix, chol1 for Cholezky decomposition with one unit impulse, comb1 for concerted action with one unit impulse.}

\item{runs}{: number of bootstrap runs to generate the confidence bands}

\item{conf}{: a two components vector of the tail probabilities of the confidence interval.}
}
\value{
An (mn,mn,nstep,3) array of IRF with columns representing the impulse rows the responses.
}
\description{
This function generates impulse response functions of an estimated GVAR with confidence bands
}
\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 2; p[,2]=1;
res_d = GVARData(m=2,n=5,p=p,T=100,type="const")
max(res_d$Y)
dim(res_d$Y)
res_e = GVARest(res = res_d)
res_e$Summary

IRF_CB = irf_GVAR_CB(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
dim(IRF_CB)
IRF_g = IRF_graph(IRF_CB,Names=NA,response=c(1,4),impulse=c(1,2,3,4), ncol=4)

}
