% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_c.R
\name{irf_MRCIGVAR}
\alias{irf_MRCIGVAR}
\title{Regime specific impulse response functions of MRCIGVAR(n,p,S)}
\usage{
irf_MRCIGVAR(
  res,
  nstep,
  comb,
  state = state,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  sigmaNPDS = NA
)
}
\arguments{
\item{res}{a list of estimated MRCIGVAR as output of MRCIGVARest or MRCIGVARestm}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{state}{an n vector specifying the speciic state for each country.}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1")}

\item{sigmaNPDS}{the state-dependent covariance matrix}
}
\value{
a list containing the impulse response functions and the accumulated impulse response function, and the boostrap parameters as well.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRCIGVAR(n,p,S).
Using the estimated G[,,,s] and Sigma[,,s] matrices of the MRGVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 3; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)

## case of n = 3, m = 2, S = 2    m: number of variables, n: number of countries


res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),r=rep(1,3),Ncommtrend=1)
max(abs(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])
res_e  = MRCIGVARestm(res=res_d)

res_e$Summary

IRF_CB = irf_MRCIGVAR_CB(res=res_e, nstep=10, comb=NA, state = c(2,1,1), irf = "gen1", runs = 20, conf = c(0.05, 0.95), NT = 1)

str(IRF_CB)

IRF_g = IRF_graph(IRF_CB[[1]])
IRFF2 = irf_MRCIGVAR(res=res_e,nstep=10,comb=NA,state=c(2,2,2),irf='gen')
plot(IRFF2[1,1,],type="l")


}
