% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{irf_MRCIVAR_CB}
\alias{irf_MRCIVAR_CB}
\title{Regime specific impulse response functions of MRCIVAR(n,p,S)}
\usage{
irf_MRCIVAR_CB(
  res_e,
  nstep = 20,
  irf = c("gen", "gen1"),
  runs = 100,
  comb = NA,
  G = NA,
  smat = NA,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res_e}{an object of MRCIVAR as output of MRVARestm}

\item{nstep}{the length of impulse response function}

\item{irf}{types of the impulse response function c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol Cholezky decomposition, chol1 Cholezky decomposition with unit impulse, comb1 concerted action with a one unit impulse.}

\item{runs}{Number of simulation runs}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{G}{The matrix used in the permanent and transitory decomposition}

\item{smat}{An explicit decomposition matrix that defines a structural shock.}

\item{conf}{a vector of the tail probabilities of the confidence interval.}
}
\value{
A list of impulse response function in two regimes and the bootstrap parameters.
}
\description{
This function calculates the regime specific impulse response functions with confidence bands, using Bo[,,,s] and Sigma[,,s] matrices of the estimated MRCIVAR.
}
\examples{

n =10

Sigma = 1:(n*n*2)
dim(Sigma) = c(n,n,2)
Sigma[,,1] = diag(n)
Sigma[,,2] = diag(n)
p=matrix(0,2,2)
p[,1] = c(3,3)
res_d = MRCIVARDatam(n=n,p=p,T=250,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=2)
#colnames(res_d$Y) = c("w","p","I","Q")
max(abs(res_d$Y))
res_e = MRCIVARestm1(res=res_d)
#res_e$Summary
if (! max(Mod(STAT(res_e$Bo[,,,1])),Mod(STAT(res_e$Bo[,,,2])) ) > 1 ) {

  IRF  = irf_MRCIVAR_CB(res_e,nstep=20,irf="gen1",runs=100,comb=NA,G=NA,conf=c(0.05,0.95))
  IRF_g1 <- IRF_graph(IRF[[1]])        # irf of regime 1
  IRF_g2<- IRF_graph(IRF[[2]])         # irf of regime 2

}

}
