% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{irf_MRVAR_NM}
\alias{irf_MRVAR_NM}
\title{Regime specific impulse response functions of MRVAR(n,p,S)}
\usage{
irf_MRVAR_NM(
  RESS,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1")
)
}
\arguments{
\item{RESS}{an output of MRVARest}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the weights used in GVAR models. For MRVAR its value is NA.}

\item{irf}{types of the impulse response functions. irf=c("gen","chol","chol1","gen1","comb1"), gen for generalized IRF with one standard deviation shocks, gen1 for generalized IRF with one unit impulse, chol for IRF with Cholezky decomposition of the covariance matrix, chol1 for Cholezky decomposition with one unit impulse, comb1 for concerted action with one unit impulse in GVAR.}
}
\value{
an (n,n,nstep,2) array of IRF with columns representing the impulse and rows the responses.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRVAR(n,p,S), using Bo[,,,s] and Sigma[,,s] matrices of the estimated MRVAR(n,p,S).
}
\examples{

p = matrix(c(2,1,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=300,S=2,SESVI=1)
max(abs(res_d$Y))
res_e  = MRVARest(res_d)
res_e$Summary

IRF    = irf_MRVAR_NM(res_e,nstep=10,comb=NA,irf="gen")
IRF_CB = irf_MRVAR_NM_CB(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.90))
IRF_list1 <-IRF_graph(IRF_CB[,,,,1])
IRF_list2 <-IRF_graph(IRF_CB[,,,,2])

}
