% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{JointX}
\alias{JointX}
\title{Join exogenous variable matrix for MRVAR}
\usage{
JointX(X)
}
\arguments{
\item{X}{The data matrix of exogenous variables}
}
\value{
An array of containing exogenous data
}
\description{
Join exogenous variable matrix for MRVAR
}
