% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{MIxCIVARData}
\alias{MIxCIVARData}
\title{Mixed VECM with I(0) and I(1) variables}
\usage{
MIxCIVARData(
  n,
  p,
  T,
  r,
  k,
  type,
  Bo = NA,
  Y = NA,
  X = NA,
  D = NA,
  Go = NA,
  B = NA,
  Sigma = NA
)
}
\arguments{
\item{n}{: dimension of the mixed CIVAR}

\item{p}{: lag of the mixed CIVAR}

\item{T}{: number of observations or length of the generated data}

\item{r}{: number of unit roots in the joint CIVAR}

\item{k}{: number of I(0) variables}

\item{type}{: types of the deterministic components in the conditional CIVAR}

\item{Bo}{: coefficient matrix of the mixed CIVAR}

\item{Y}{: data of the mixed CIVAR}

\item{X}{: data of the exogenous variables}

\item{D}{: transformation matrix to mix the I(0) and I(1) components}

\item{Go}{: pre-loaded selection matrix}

\item{B}{: coefficient matrix of the mixed CIVAR}

\item{Sigma}{: covariance matrix of the residuals}
}
\value{
a list contains estimation and test results of the mixed VECM
}
\description{
This function generates and estimates a mixed VECM with I(0) and I(1) variables
}
\examples{

#RR <- MIxCIVARData(n=9,p=2,T=209,r=5,k=2,type="const",Bo=NA,Y=NA,D=NA)
## DGP of I(1) and I(0) mixed VECM
#plot(ts(RR$Y))

#### testing the mixed VECM via testing the restrictions on beta
#res_d <- CIVARData(n=9,p=2,T=209,type="const",crk=4)
#res_e = CIVARest(res=res_d)
#res_e$Summary
#n = 9; crk = 4; k = 2; r = 5
#CC  <- c(8,9,17,18)
#GG  <- c(19:25,28:34)
#G = diag(n*crk); psi=matrix(1,n*crk,1)
#### this implies there is no restrictions on the adjustment coefficients alpha
#H = diag(n*crk);              H2 = H[,-c(seq(1,(n-r-k)*n,n),seq((n-r-k+1)*n,(n-r)*n,n),CC,GG)]
#### only normalization
#h = matrix(0,n*crk,1);         h[c(seq(1,(n-r-k)*n,n),seq((n-r-k+1)*n,(n-r)*n,n)),1] <- 1
#phi = matrix(1,ncol(H2),1)
### check consistency of the restrictions    G\%*\%psi; H2\%*\%phi + h
#
#G\%*\%psi; H2\%*\%phi + h
#
#ABtest = CIVARTest(res=res_d,H=H2,h=h,phi=phi,G=G,Dxflag=0)
#ABtest$betar
#ABtest$alphar
#ABtest$VECMR$coefficients
#ABtest$LR
#1-pchisq(ABtest$LR,14)   ### The Ho of last two are I(0) can be rejected
#RR$GABtest$LR
#
}
