% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_d.R
\name{MRCIGVAR_Select}
\alias{MRCIGVAR_Select}
\title{Calculation of information criteria AIC and BIC for an estimated MRCIGVAR model}
\usage{
MRCIGVAR_Select(res, L_V = L_V, TH_V = TH_V)
}
\arguments{
\item{res}{: an MRCIGVAR object generated from MRCIGVARData or estimated from MRCIGVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag length and the foreign lag length respectively.}

\item{TH_V}{: a vector containing possible threshold values .}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a matrix with different lag specifications,  threshold values and the corresponding information criteria.
}\if{html}{\out{</div>}}
}
\description{
Calculation of information criteria AIC and BIC for an estimated MRCIGVAR model
}
\examples{

m = 2                  ## m: number of variables, n: number of countries
n = 4
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:4,,]; p[,1:2,2] = 3

TH = c(1:4)*0; dim(TH) = c(1,4)

res_d <- MRCIGVARData(m=2,n=4,p=p,TH=TH,T=400,S=2, SESVI=c(1,3,5,7),r=rep(1,4))

res_em = MRCIGVARest(res=res_d)

L_v = c(3,3)
TH_v = c(0,0.1)
res_d$p

MRCIGVARSelect = MRCIGVAR_Select(res=res_d,L_V=L_v,TH_V=TH_v)
dim(MRCIGVARSelect)
MRCIGVARSelect[which.min(MRCIGVARSelect[,28]),]

}
