% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_c.R
\name{MRCIGVARData}
\alias{MRCIGVARData}
\title{Data generating process of MRCIGVAR(m,n,p,S,r)}
\usage{
MRCIGVARData(
  m,
  n,
  p,
  T,
  S,
  W = NA,
  SESVI = NA,
  TH = NA,
  Go = NA,
  Ao = NA,
  Bo = NA,
  Sigmao = NA,
  Uo = NA,
  SV = NA,
  type = NA,
  Co = NA,
  X = NA,
  Yo = NA,
  d = NA,
  r = NA,
  r_np = NA,
  Ncommtrend = NA
)
}
\arguments{
\item{m}{: number of variables in each country/unit}

\item{n}{: number of countries/units}

\item{p}{: an (n, 3, S) array, each raw contains the lag length of the domestic variables, the lag length of the foreign variables and the number of the exogenous variables of the corresponding country in the respective regimes.}

\item{T}{: number of observations.}

\item{S}{: number of regimes}

\item{W}{: an (n x n) weighting matrix. w_ij is the weight of country j in the foreign variables of i-th country diag(W)=0}

\item{SESVI}{: an n-vector of indices of the switching variables across n countries. Eg SESVI = seq(1,m*n,m).}

\item{TH}{: an (n x S-1) matrix of threshold values}

\item{Go}{: an (mn,mn,p,S) array containing coefficients of MRCIGVAR. Go is constructed from Bo, Ao and W.}

\item{Ao}{: an (m, m, p, n, S) array containing the coefficients of foreign variables}

\item{Bo}{: an (m, m, p, n, S) array containing the coefficients of domestic variables.}

\item{Sigmao}{: an (mn, mn, S) array of the covariance matrix of MRGVAR(m,n,p,S)}

\item{Uo}{: a (T, mn, S) array of the temporally independent innovations}

\item{SV}{: exogenous switching variables}

\item{type}{: types of the deterministic component. "const", "none", "exog0", and "exog1" are 4 options}

\item{Co}{: an (m , k+1, n, S) array containing the coefficients of the deterministic components of the n countries.}

\item{X}{: a (T x k x n x S) array of exogenous variables.}

\item{Yo}{: initial values}

\item{d}{: the time lag between signal and regime-switching}

\item{r}{: an n-vector containing the number of unit root process in each country}

\item{r_np}{: an (m, Pmax, n, S) array containing the roots of the characteristic polynomials of each country.}

\item{Ncommtrend}{: number of common stochastic trends in the MRCIGVAR.}
}
\value{
\preformatted{ an MRGVAR object containing the generated data, the used parameters and the exogenous variables.
}

\itemize{
\item Y     : a (T x mn)  matrix of simulated data
\item X     : a (T x k x n x S) array of exogenous variables
\item Uo    : a (T, mn,S) array of the simulated innovations of the MRGVAR(m,n,p,S)
\item C     : an (nm, (k+1),S) array containing the coefficients of the deterministic components.
\item St    : a (T x n) matrix of the simulated time path of states/regimes
\item check : maximum of the data for checking the stationarity
}
}
\description{
This function generates data from an multi-regime cointegrated global VAR process and returns an MRCIGVAR(m,n,p,S) object that is a list containing the data and the parameters used in the MRCIGVAR(m,n,p,S) process.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 3; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI = c(1,3,5)
r  = rep(1,n)

## case of n = 3, m = 2, S = 2   with m: number of variables, n: number of countries.

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),r=rep(1,3),Ncommtrend=1)
max(abs(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])
res_e  = MRCIGVARestm(res=res_d)


T= 300
XX = matrix(rnorm(T*6*2),T*6,2)
dim(XX) = c(T,2,3,2)

XX = XX*10
p[,3,]=2

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=T,S=2, SESVI=c(1,3,5),type="exog0",r=rep(1,3),Ncommtrend=1,X=XX)

max(abs(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])

res_e  = MRCIGVARestm(res=res_d)

res_e$Summary

IRF_CB = irf_MRCIGVAR_CB(res=res_e, nstep=10, comb=NA, state = c(2,1,1), irf = "gen1", runs = 20, conf = c(0.05, 0.95), NT = 1)

str(IRF_CB)

IRF_g = IRF_graph(IRF_CB[[1]])

}
