% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_d.R
\name{MRCIGVARDataR}
\alias{MRCIGVARDataR}
\title{Data generating process of MRCIGVARDataR(res)}
\usage{
MRCIGVARDataR(res)
}
\arguments{
\item{res}{: an output of MRCIGVARest}
}
\value{
: an MRCIGVAR object.
}
\description{
This function will generate data from a MRCIGVAR object. It will generate enough data for estimation purpose.
}
