% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_c.R
\name{MRCIGVARest}
\alias{MRCIGVARest}
\title{Estimation of MRCIGVAR(m,n,p,S,r)}
\usage{
MRCIGVARest(res)
}
\arguments{
\item{res}{: an MRCIVAR object of the output of MRCIGVARData}
}
\value{
\preformatted{ : an MRCIVAR object with estimated parameters and statistics.
}
}
\description{
This function estimates the parameters of a multi regime cointegrated global VAR model. The adjustment speeds are assumed to be identical in different regimes.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 3; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)

## case of n = 3, m = 2, S = 2     ## m: number of variables, n: number of countries

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),r=rep(1,3),Ncommtrend=1)
max(abs(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])
max(abs(res_d$Y))
res_e  = MRCIGVARest(res=res_d)



}
