% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{MRCIVAR_Selectm}
\alias{MRCIVAR_Selectm}
\title{Calculation of information criteria AIC and BIC for MRCIVAR models}
\usage{
MRCIVAR_Selectm(res = res, L_V = L_V, TH_V = TH_V)
}
\arguments{
\item{res}{: an MRCIVAR object generated from MRCIVARDatam or estimated from MRCIVARestm1.}

\item{L_V}{: a 2 components vector containing the maxima of the lags length of the two regimes, respectively.}

\item{TH_V}{: a vector containing possible threshold values for the selection.}
}
\value{
\preformatted{ : a matrix with different lag specifications, threshold values, and the corresponding model selection criteria.
}
}
\description{
Calculation of information criteria AIC and BIC for MRCIVAR models
}
\examples{

Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2)
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)

res_d = MRCIVARDatam(n=4,p=p,T=2610,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=1)
res_e = MRCIVARestm1(res=res_d)

TH_v = c(0,0.1)
L_v = c(6,6)

Selm = MRCIVAR_Selectm(res=res_e,L_V=L_v,TH_V=TH_v)
MRVAR_Select_Summary(Selm)

}
