% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{MRCIVARDatam}
\alias{MRCIVARDatam}
\title{Data generating process of MRCIVAR(n,p,S)}
\usage{
MRCIVARDatam(
  n = 2,
  p = matrix(2, 2, 2),
  T = 100,
  S = 2,
  SESVI,
  TH,
  Bo,
  Co,
  Sigmao,
  Uo,
  SV,
  type,
  X,
  mu,
  Yo,
  Do,
  d = 1,
  r = 1
)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: an (S x 2) matrix. Each row of p contains the lag length for the corresponding state and the number of exogenous variables for the state.}

\item{T}{: number of observations}

\item{S}{: number of regimes}

\item{SESVI}{: index of the switching variable, switching sv_t = Y[t-1,SESVI] > Y[t-2,SESVI]\preformatted{           (n,S,p,T,SESVI) are parameters that have to be provided.
}}

\item{TH}{: an (S-1) vector of threshold values}

\item{Bo}{: an (n,n,p,S) array of the coefficients  of MRVAR(n,p,S). If not given it will be generated.}

\item{Co}{: an (n,k+1,S) array of the coefficients of the deterministic components. For type="none" Co = O*(1:n,1:S), for "const" Co is an n-vector for each regime, "exog0" Co is a (n,k+1,S) array with first column of zeros for each regime respectively,for "exog1" Co is an (n,1+k, S) array without zero restrictions.}

\item{Sigmao}{: an (n,n,S) array containing S covariance matrices of the residuals}

\item{Uo}{: residuals, if it is not NA it will be used as input to generate the MRVAR(n,p,S) for each regime.}

\item{SV}{: exogenous switching variable}

\item{type}{: type of the deterministic components type = ("none","const","exog0","exog1")}

\item{X}{: a (T x k x S) matrix of exogenous variables for each state. The second dimension can be filled with zeros to take into account that the the exogenous variables are not identical in each state.}

\item{mu}{: an (n x S) matrix of the regime specific mean of the variables}

\item{Yo}{: a (p, n, S) array of initial values of the process}

\item{Do}{: a (T, n, S) array of extra exogenous components (not used with value zero)}

\item{d}{: lag delay of the self-exiting switching}

\item{r}{: number of unit roots in the system. (n - r) is then the cointegration rank.\preformatted{          (TH,Bo,Sigmao,Uo,SV) if not provided, they will be generated randomly.
}}
}
\value{
\preformatted{ An MRCIVAR object containing the generated data, the used parameters, and the exogenous variables.
}
}
\description{
This function generates data from an multi-regime cointegrated VAR(p) process with identical cointegration relations and different adjustment speeds and returns a MRCIVAR which is a list containing data and parameters used in the multi regime cointegrated VAR(p) process.
}
\examples{
Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2)
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,3)
res_d = MRCIVARDatam(n=4,p=p,T=250,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=2)
colnames(res_d$Y) = c("w","p","I","Q")
max(abs(res_d$Y))
res_e = MRCIVARestm1(res=res_d)
res_e$Summary

p=matrix(0,2,2)
p[,1] = c(3,3)
p[,2] = 1
T = 200
XX = matrix(rnorm(T),T,1)
XX = cbind(XX,XX); dim(XX) = c(T,1,2)
dim(XX)

res_d = MRCIVARDatam(n=4,p=p,T=T,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="exog0",r=1,X=XX)
res_e = MRCIVARestm1(res=res_d)
res_e$Summary

res_d = MRCIVARDatam(n=4,p=p,T=T,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="exog1",r=1,X=XX)
res_e = MRCIVARestm1(res=res_d)
res_e$Summary

}
