% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{MRCIVARestm1}
\alias{MRCIVARestm1}
\title{Estimation of MRCIVAR models}
\usage{
MRCIVARestm1(res)
}
\arguments{
\item{res}{an MRCIVAR object of the output of MRCIVARDatam}
}
\value{
an estimated MRCIVAR object containing estimated parameters and test results
}
\description{
This function executes a Johansen test of cointegration ranks and estimates the parameters of a MRCIVAR(n,p,S) model.
}
\examples{
Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2)
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,3)

res_d = MRCIVARDatam(n=4,p=p,T=250,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=2)
colnames(res_d$Y) = c("w","p","I","Q")
max(abs(res_d$Y))
res_e = MRCIVARestm1(res=res_d)
res_e$Summary

}
