% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{MRCVECMest2}
\alias{MRCVECMest2}
\title{Estimation of a multi regime conditional vector error correction process.}
\usage{
MRCVECMest2(
  y,
  x,
  s,
  model = c("I", "II", "III", "IV", "V"),
  type = c("eigen", "trace"),
  crk = 2,
  P = matrix(2, 2, 2),
  q = 0.95,
  Dxflag = 0,
  X = NA,
  CZ = NA
)
}
\arguments{
\item{y}{: data matrix of the endogenous variable}

\item{x}{: x data matrix of the conditioning variables. If x is missing, it will estimate a unconditional VECM.}

\item{s}{: the series of state variable of value (0,1) for the two regime case. Missing s implies single regime VECM.}

\item{model}{: It assumes one of the values in c("I","II","III","IV","V") corresponding to the five cases in Johansen test.}

\item{type}{: c("eigen", "trace") corresponding to the trace test or the max eigenvalue test}

\item{crk}{: cointegration rank}

\item{P}{: 2 x 2 matrix containing the lag of the cointegrated VAR process. The first row is the lag of the first regime and the number of exogenous variables. If the second row is zero, this is a one regime VECM.}

\item{q}{: The significance level}

\item{Dxflag}{: A flag indicating if the conditioning variables enter the cointegration space}

\item{X}{: The conditioning variables}

\item{CZ}{: Common factor variables in CIGVAR and MRCIGVAR}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A list containing: the result of the Johansen test, VECM in regression format, lambda, beta, PI, GAMMA, model, P, s
}\if{html}{\out{</div>}}
}
\description{
This function estimates the unknown parameters of a multi regime conditional VECM based on provided data.
}
