% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{MRCVECMestm}
\alias{MRCVECMestm}
\title{Estimation of a two regime conditional vector error correction model.}
\usage{
MRCVECMestm(
  y,
  x,
  s,
  model = c("I", "II", "III", "IV", "V"),
  type = c("eigen", "trace"),
  crk = 2,
  P = matrix(2, 2, 2),
  q = 0.95,
  Dxflag = 0,
  X = X,
  CZ = NA
)
}
\arguments{
\item{y}{: data matrix of the endogenous variables}

\item{x}{: data matrix of the conditioning variables. If x is missing, it will estimate an unconditional VECM.}

\item{s}{: the series of state variable with values (0,1) for the two regimes. Missing s implies single regime VECM.}

\item{model}{: It assumes one of the values in c("I","II","III","IV","V") corresponding to the five cases in the Johansen test.}

\item{type}{: c("eigen", "trace") corresponding to the trace test or the max eigenvalue test}

\item{crk}{: cointegration rank}

\item{P}{: a (2 x 2) matrix containing the lag of the cointegrated VAR process. The first row is the lag of the first regime and the number of exogenous variables. If the second row is zero, this is a one regime VECM.}

\item{q}{: The significance level used}

\item{Dxflag}{: a flag indicating if the conditioning variables enter the cointegration space}

\item{X}{: the exogenous stationary variables}

\item{CZ}{: the exogenous non-stationary common factors}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ : a list containing: the result of JH test, VECM in regression format, lambda, beta, PI, GAMMA, model, P, s
}\if{html}{\out{</div>}}
}
\description{
This function estimates the unknown parameters of a two regime conditional VECM based on provided data. The cointegrating vectors are identical in the two regimes but the adjustment speeds are different in the two regimes.
}
