% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{MRGVARData}
\alias{MRGVARData}
\title{Data generating process of MRGVAR(m,n,p,S)}
\usage{
MRGVARData(
  m,
  n,
  p,
  T,
  S,
  W = NA,
  SESVI = NA,
  TH = NA,
  Go = NA,
  Ao = NA,
  Bo = NA,
  Sigmao = NA,
  Uo = NA,
  SV = NA,
  type = NA,
  Co = NA,
  X = NA,
  Yo = NA,
  d = NA
)
}
\arguments{
\item{m}{: number of variables in a country/unit}

\item{n}{: number of countries/units}

\item{p}{: an (n, 3, S) array, each raw specifies the lag length of the domestic variables, the lag length of the foreign variables and the number of exogenous variables for the respective regime.}

\item{T}{: number of observations.}

\item{S}{: number of regimes.}

\item{W}{: an (n x n) weighting matrix. w_ij is the weight of country j in the foreign variables of i-th country diag(W)=0}

\item{SESVI}{: an n-vector of indices of the switching variables across n countries. Eg. SESVI = seq(1,m*n,m).}

\item{TH}{: an (n, S-1) matrix of threshold values}

\item{Go}{: an (mn,mn,p,S) array of the MRGVAR(m,n,p,S) coefficients. G is constructed from Bo, Ao and W.}

\item{Ao}{: an (m, m, p, n, S) array collecting coefficients of foreign variables}

\item{Bo}{: an (m, m, p, n, S) array collecting coefficients of domestic variables}

\item{Sigmao}{: an (mn, mn, S) array of the covariance matrix of MRGVAR(m,n,p,S)}

\item{Uo}{: a (T, mn, S) array of the temporally independent innovation processes}

\item{SV}{: exogenous switching variables}

\item{type}{: types of deterministic component "const", "none", "exog0", and  "exog1" are 4 options}

\item{Co}{: an (m , k+1, n, S) array collecting the coefficients of the deterministic components of the n countries for every regime.}

\item{X}{: a (T x k x n x S) matrix of exogenous variables.}

\item{Yo}{: Initial values}

\item{d}{: the time lag between signal and switching}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a MRGVAR object containing the generated data, the used parameters and the exogenous variables.
}\if{html}{\out{</div>}}

res_d = list(Y,X,Uo,resid,Go,GDC,Co,Sigmao,TH,St,SV,SESVI,Ao,Bo,check,type,m,n,p,S,W,SigmaS,Yo,d)
\itemize{
\item Y     : a (T x nm)  matrix of simulated data via of the MRGVAR(m,n,p,S)
\item X     : a (T x k x n x S) matrix of exogenous variables.
\item Uo    : a (T, mn,S) array of the simulated innovations of the MRGVAR(m,n,p,S)
\item C     : an (nm, (k+1),S) array containing the coefficients of the deterministic components.
\item St    : a (T x n) matrix of the simulated time path of states/regimes
\item check : maximum of the data for checking the stationarity
}
}
\description{
This function generates data from an multi-regime stationary global VAR(p) process and returns an MRGVAR object that is a list containing the generated data and the used parameters.
}
\examples{
## case of n = 2, m = 2, S = 2     ## m: number of variables, n: number of countries
p = rep(1,12); dim(p) = c(2,3,2)
p[1,1,2] = 2; p[2,2,2]=2; p[,3,] = 0
TH = c(1:2)*0; dim(TH) = c(1,2)
res_d <- MRGVARData(m=2,n=2,p=p,TH=TH,T=100,S=2,SESVI=c(1,3),type="const")
max(res_d$Y)

### estimation of the MRGVAR model
colnames(res_d$Y) = c("P","Q","Pa","Qa")
res_e = MRGVARest(res=res_d)
res_e$Summary

IRF_CB  = irf_MRGVAR_CB(res=res_e,nstep=10,comb=NA,state=c(1,1),irf="gen1",runs=20,conf=c(0.05,0.95))
IRF_g = IRF_graph(IRF_CB[[1]],Names=c("P","Q","Pa","Qa"))    #IRF
#IRF_g = IRF_graph(IRF_CB[[2]])   # accumulated IRF
}
