% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{MRGVARDataR}
\alias{MRGVARDataR}
\title{Data generating process of MRGVARDataR(res)}
\usage{
MRGVARDataR(res)
}
\arguments{
\item{res}{: an output of MRGVARest}
}
\value{
: an MRGVAR object.
}
\description{
This function will generate data from a MRGVAR object. It will generate enough data for estimation purpose.
}
