% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{MRGVARest}
\alias{MRGVARest}
\title{Estimation of MRGVAR(m,n,p,S) models}
\usage{
MRGVARest(res)
}
\arguments{
\item{res}{: an MRGVAR object which is an output of MRGVARData.}
}
\value{
res  : an MRGVAR object with estimated parameters and test statistics.
}
\description{
This function estimates the parameters of a specified MRGVAR(m,n,p,S) model based on provided data.
}
\examples{

## case of n = 2, m = 2, S = 2     ## m: number of variables, n: number of countries
p = rep(1,12); dim(p) = c(2,3,2)
p[1,1,2] = 2; p[2,2,2]=2; p[,3,] = 0
TH = c(1:2)*0; dim(TH) = c(1,2)
res_d <- MRGVARData(m=2,n=2,p=p,TH=TH,T=100,S=2,SESVI=c(1,3),type="const")
max(res_d$Y)

### estimation of the MRGVAR model
colnames(res_d$Y) = c("P","Q","Pa","Qa")
res_e = MRGVARest(res=res_d)
res_e$Summary

IRF_CB  = irf_MRGVAR_CB(res=res_e,nstep=10,comb=NA,state=c(1,1),irf="gen1",runs=20,conf=c(0.05,0.95))
IRF_g = IRF_graph(IRF_CB[[1]],Names=c("P","Q","Pa","Qa"))    #IRF
#IRF_g = IRF_graph(IRF_CB[[2]])   # accumulated IRF
}
