% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{MRVAR_Select_Summary}
\alias{MRVAR_Select_Summary}
\title{Summary result of MRVAR model selection}
\usage{
MRVAR_Select_Summary(Sel)
}
\arguments{
\item{Sel}{An output of MRVAR_Select}
}
\value{
The optimal model according to BIC or AIC
}
\description{
Summary result of MRVAR model selection
}
