% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{MRVARData}
\alias{MRVARData}
\title{Data generating process of MRVAR(n,p,S)}
\usage{
MRVARData(
  n,
  p,
  T,
  S,
  SESVI,
  TH,
  Bo,
  Co,
  Sigmao,
  Uo,
  SV,
  type,
  X,
  mu,
  Yo,
  Do,
  d
)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: an (S x 2) matrix. Each row of p specifies the lag length and the number of exogenous variables for the corresponding regime.}

\item{T}{: number of observations}

\item{S}{: number of regimes}

\item{SESVI}{: index of the switching variable\preformatted{           (n,S,p,T,SESVI) must be provided.
}}

\item{TH}{: an (S-1)-vector of threshold values}

\item{Bo}{: an (n,n,p,S) array of the coefficients of the MRVAR(n,p,S) model. If Bo is not given it will be generated.}

\item{Co}{: an (n,k+1,S) array of the coefficients of the deterministic components. For type="none" Co = O*(1:n,1:S), for "const" Co is an (1:n,1:S) array, for "exog0" Co is an (n,k+1,S) array with first column zero for each regime respectively,for "exog1" Co is an (n,1+k, S) array without zero restrictions.}

\item{Sigmao}{: an (n,n,S)   array containing S covariance matrices of residuals, each for one regime.}

\item{Uo}{: residuals, if it is not NA it will be used as input to generate data of the MRVAR(n,p,S) process.}

\item{SV}{: exogenous switching variable}

\item{type}{: type of the deterministic components type = c("none","const","exog0","exog1")}

\item{X}{: a (T x k x S) array of exogenous variables for each state. The second dimension can be filled with zeros to take into account that the the exogenous variables are not identical in each state.}

\item{mu}{: an (n x S) matrix of the regime specific means of the variables}

\item{Yo}{: a (p, n, S) array of initial values of the process}

\item{Do}{: a (T, n, S) array of extra exogenous components (not used with value zero)}

\item{d}{: lag delays of the self-exiting switching variable.\preformatted{         (TH,Bo,Co,Sigmao,Uo,SV,type) if not provided, they will be generated randomly.
}}
}
\value{
\preformatted{ an MRVAR(n,p,S) object that is a list containing the generated data, the used parameters and the exogenous variables.
}
}
\description{
This function generates data from an multi-regime stationary VAR(p) process and returns an MRVAR(n,p,S) object that is a list containing data and the parameters used in the MRVAR(n,p,S) process.
}
\examples{
p = matrix(c(3,3,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=200,S=2,SESVI=1,type="const")
colnames(res_d$Y) = c("Y","P")
STAT(res_d$B[,,,1])
STAT(res_d$B[,,,2])
res_e = MRVARest(res=res_d)
res_e$Summary
}
