% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{MRVARest}
\alias{MRVARest}
\title{Estimation of MRVAR(n,p,S)}
\usage{
MRVARest(res)
}
\arguments{
\item{res}{:an object of MRVAR that is an output of MRVARData including at least: n, S, p, type, Y, SESVI, TH, d, and optionally X.}
}
\value{
an MRVAR object that is a list containing estimated parameters and some test statistics.
}
\description{
This function estimates the parameters of a specified MRVAR(n,p,S) based on provided data.
}
\examples{
p = matrix(c(3,3,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=200,S=2,SESVI=1,type="const")
colnames(res_d$Y) = c("Y","P")
STAT(res_d$B[,,,1])
STAT(res_d$B[,,,2])
res_e = MRVARest(res=res_d)
res_e$Summary

}
