% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_G}
\alias{p_G}
\title{Weighting matrix for GVAR based in GDP}
\format{
An object of class \code{array} of dimension 15 x 3 x 2.
}
\usage{
p_G
}
\description{
Weighting matrix for GVAR based in GDP
}
\keyword{datasets}
