% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{Redex}
\alias{Redex}
\title{Re-indexing the order of the endogenous and exogenous variables}
\usage{
Redex(N1, N2, p)
}
\arguments{
\item{N1}{dimension of the endogenous variables}

\item{N2}{dimension of the exogenous variables}

\item{p}{lag length}
}
\value{
a new index
}
\description{
Re-indexing the order of the endogenous and exogenous variables
}
