% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{rnormSIGMA_cond}
\alias{rnormSIGMA_cond}
\title{Conditional normal random numbers}
\usage{
rnormSIGMA_cond(T, sigma, I, v, switch)
}
\arguments{
\item{T}{Length of generated time series}

\item{sigma}{The (n x n) covariance matrix of the normal series}

\item{I}{Index of conditioning component}

\item{v}{The value of the conditioning component}

\item{switch}{A switch variable: switch = 1 gives the conditional random series and switch = 0 gives the expected values.}
}
\value{
A (T x (n-1)) matrix of iid conditional normal time series or the conditional expected values
}
\description{
This function generates random numbers from iid multivariate conditional normal distribution with covariance matrix Sigma, given i-th component has the value of v, this will be an (n-1) dimensional random number
}
