% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_c.R
\name{Roots2coef}
\alias{Roots2coef}
\title{Roots2Coefficients}
\usage{
Roots2coef(p, r_p)
}
\arguments{
\item{p}{The lag length}

\item{r_p}{A p-vector of roots outside the unit circle}
}
\value{
A vector of AR coefficients
}
\description{
Given the roots of a characteristic polynom in lags, output the coefficients of the corresponding AR process.
}
\examples{

Roots2coef(3,c(1.1,1.2,1.3))
}
