% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{RZSt2VECM}
\alias{RZSt2VECM}
\title{Calculation of restricted MRCIVAR parameters from data moment matrices}
\usage{
RZSt2VECM(R0, R1, Y0, Z2, Z1, St, crk)
}
\arguments{
\item{R0}{I(0) residuals of the auxiliary regression}

\item{R1}{I(0) residuals of the auxiliary regression}

\item{Y0}{The first difference series of the endogenous I(1) variables}

\item{Z2}{The first difference series with lags}

\item{Z1}{The level series}

\item{St}{The regime indicator series}

\item{crk}{The cointegration rank}
}
\value{
a list containing estimated constrained parameters as initial values for further iterations
}
\description{
Calculation of restricted MRCIVAR parameters from data moment matrices
}
