% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_d.R
\name{ShiftZ2}
\alias{ShiftZ2}
\title{Shift of Z2 regressors}
\usage{
ShiftZ2(Z2, kz, n1, p)
}
\arguments{
\item{Z2}{: Regressor's matrix to be shifted}

\item{kz}{: number of common exogenous factors}

\item{n1}{: number of domestic variables}

\item{p}{: the lag length of domestic variables in Z2 which is also the lag length of exogenous common factors}
}
\value{
a shifted Z2 if there are exogenous common factors. It returns Z2 without shift if there are no common factors.
}
\description{
This function generates impulse response functions of an estimated CIGVAR(m,n,p) with confidence bands.
}
