% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{SigmaNPD}
\alias{SigmaNPD}
\title{Calculate a state dependent covariance matrix from GVAR residuals}
\usage{
SigmaNPD(res, StateT)
}
\arguments{
\item{res}{An estimated GVAR object}

\item{StateT}{A vector of selected states of each country}
}
\value{
The covariance matrix of the selected states
}
\description{
Calculate a state dependent covariance matrix from GVAR residuals
}
