% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVAR_c.R
\name{SigmaNPDSelectR}
\alias{SigmaNPDSelectR}
\title{Select a state-dependent covariance matrix from a GVAR model}
\usage{
SigmaNPDSelectR(res, StateT)
}
\arguments{
\item{res}{An estimated GVAR object}

\item{StateT}{A vector of selected regimes of each country}
}
\value{
A positive definite covariance matrix of the selected states each country
}
\description{
Select a state-dependent covariance matrix from a GVAR model
}
