% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{STAT}
\alias{STAT}
\title{Root of the characteristic polynomial in Lag}
\usage{
STAT(G)
}
\arguments{
\item{G}{a coefficient matrix of dimension (n x n x p) of an VAR(p) object.}
}
\value{
Roots of the characteristic polynomial of VAR(p)
}
\description{
Root of the characteristic polynomial in Lag
}
