% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{summaryCIVAR}
\alias{summaryCIVAR}
\title{This function replaces the variable names in an lm output.}
\usage{
summaryCIVAR(lm1 = lm1, sname = "Z2\\\\[,4:9\\\\]")
}
\arguments{
\item{lm1}{The coefficient matrix of the output of lm}

\item{sname}{The elements in column names to be replaced}
}
\description{
This function replaces the variable names in an lm output.
}
