% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_c.R
\name{SW2comb}
\alias{SW2comb}
\title{Transformation of the weighting vector to a weighting matrix}
\usage{
SW2comb(SW, n, N, K)
}
\arguments{
\item{SW}{The weighting vector}

\item{n}{The number of variables}

\item{N}{The number of countries}

\item{K}{The number}
}
\value{
The weighting matrix
}
\description{
Transformation of the weighting vector to a weighting matrix
}
