% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tab5}
\alias{Tab5}
\title{Critical values from David Mackinnon}
\format{
An object of class \code{data.frame} with 24 rows and 10 columns.
}
\usage{
Tab5
}
\description{
Critical values from David Mackinnon
}
\keyword{datasets}
