% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TH_G}
\alias{TH_G}
\title{Threshold values determined by information criterion}
\format{
An object of class \code{numeric} of length 15.
}
\usage{
TH_G
}
\description{
Threshold values determined by information criterion
}
\keyword{datasets}
