% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVAR_c.R
\name{U1}
\alias{U1}
\title{Initial U1 for DATA generating IRF}
\usage{
U1(
  B,
  sigma,
  irf = c("gen", "chol", "chol1", "gen1", "genN1", "comb", "comb1", "smat",
    "concerts1", "concerts0", "concertc"),
  comb,
  G = NA,
  smat = NA
)
}
\arguments{
\item{B}{The coefficient matrix of a VAR object}

\item{sigma}{The covariance matrix of a VAR object}

\item{irf}{Types of impulse response functions}

\item{comb}{Types of concerted policy actions}

\item{G}{The matrix used in the permanent and transitory decomposition}

\item{smat}{An explicit decomposition matrix that defines a structural shock.}
}
\value{
The transformation matrix for shock decomposition
}
\description{
Initial U1 for DATA generating IRF
}
