% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_d.R
\name{VARBS_commtrend}
\alias{VARBS_commtrend}
\title{Generating parameter matrices with common trends across countries}
\usage{
VARBS_commtrend(m, p, T, r, Ncommtrend, n, S, r_npo)
}
\arguments{
\item{m}{Number of variables of a country}

\item{p}{Lags}

\item{T}{Number of observations}

\item{r}{Number of stochastic trends}

\item{Ncommtrend}{Number of common trends}

\item{n}{Number of countries in the GVAR model}

\item{S}{Number of regimes}

\item{r_npo}{A matrix specifying the  roots of the characteristic polynomial in lags}
}
\value{
A list contain VAR parameters B, alpha, and beta
}
\description{
Generating parameter matrices with common trends across countries
}
