% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_c.R
\name{VARest}
\alias{VARest}
\title{Estimation of VAR(p)}
\usage{
VARest(res)
}
\arguments{
\item{res}{:a VAR(p) object which is an output of VARData including at least: n, p, type, Y and optionally X and type.}
}
\value{
res  :a VAR(p) containing estimated parameter values, AIC, BIC and LH.
}
\description{
This function estimates parameters of a specified VAR(p) model based on provided data.
}
\examples{
res_d = VARData(n=2,p=2,T=100,type="const")
res_e = VARest(res=res_d);
IRF_CB = irf_VAR_CB(res=res_e,nstep=20, comb=NA, irf = "gen1", runs = 100, conf = c(0.05, 0.95))
IRF_list <-IRF_graph(IRF_CB)
res_e$Summary

}
