% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_c.r
\name{VECM2VAR}
\alias{VECM2VAR}
\title{Transformation of the estimated parameters of a VECM into the parameters of the corresponging VAR}
\usage{
VECM2VAR(param, beta, q = c(1, 2, 2, 2, 2), s = NA)
}
\arguments{
\item{param}{The parameter of an estimated VECM}

\item{beta}{The cointegration vectors}

\item{q}{A vector specifying different type of VECMs}

\item{s}{The indicator of multi regime VECMs}
}
\value{
A list of (A, B, C) with B the auto regression parameter matrix, C the parameter of the deterministic components, and A the parameter matrix of foreign variables for CIGVAR.
}
\description{
Transformation of the estimated parameters of a VECM into the parameters of the corresponging VAR
}
