% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR_c.R
\name{VECM2VARm}
\alias{VECM2VARm}
\title{Transformation of parameters of MRVECM to paramters of MRCIVAR}
\usage{
VECM2VARm(param, beta, p = c(1, 2, 2, 2, 2), s = NA)
}
\arguments{
\item{param}{The coefficient matrix of the estimated multi regime vector error correction model}

\item{beta}{The estimated cointegration vectors}

\item{p}{The lags of the MRCIVAR model}

\item{s}{Types of MRCIVAR models}
}
\value{
The parameter of the multi regime cointegrated VAR
}
\description{
Transformation of parameters of MRVECM to paramters of MRCIVAR
}
