% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wmat_G}
\alias{Wmat_G}
\title{Weighting matrix for GVAR based in GDP}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 15 rows and 15 columns.
}
\usage{
Wmat_G
}
\description{
Weighting matrix for GVAR based in GDP
}
\keyword{datasets}
