% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_d.r
\name{xABF}
\alias{xABF}
\title{This function solves AB_SVAR from the reduced form and return the A, B matrices and the sum of squared errors}
\usage{
xABF(x, A0, B0, Sigma)
}
\arguments{
\item{x}{: difference between the reduced form and the AB form}

\item{A0}{: A matrix in an AB_SVAR model}

\item{B0}{: B matrix in an AB_SVAR model}

\item{Sigma}{: Covariance matrix of the reduced form}
}
\value{
difference
}
\description{
This function solves AB_SVAR from the reduced form and return the A, B matrices and the sum of squared errors
}
