% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_c.r
\name{B2CIB}
\alias{B2CIB}
\title{Transformation of the coefficient matrix of a CIGVAR to the coefficient matrix of the corresponding VECM.}
\usage{
B2CIB(B)
}
\arguments{
\item{B}{an (m,m,L) array containing the coefficients of the level VAR.}
}
\value{
\preformatted{ A list containing three components.
}

\itemize{
\item CIB        : the coefficients matrix of the error correction form
\item alpha	     : the adjustment coefficients
\item beta       : the cointegration vectors
\item sm         : the discrepancy between the selected alpha beta and CIB.
}
}
\description{
Transformation of the coefficient matrix of a CIGVAR to the coefficient matrix of the corresponding VECM.
}
