% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR_d.R
\name{CIGVARData}
\alias{CIGVARData}
\title{Data generating process of CIGVAR(m,n,p)}
\usage{
CIGVARData(
  m,
  n,
  p,
  T,
  W = NA,
  r_npo = NA,
  Ao = NA,
  Bo = NA,
  Co = NA,
  Uo = NA,
  Sigmao = NA,
  type = NA,
  X = NA,
  mu = NA,
  DFYflag = NA,
  crk = NA,
  Ncommfakt = 1,
  A = NA,
  uz = NA
)
}
\arguments{
\item{m}{: number of variables in each country/unit}

\item{n}{: number of countries/units}

\item{p}{: an (n x 3) matrix, each raw specifies the lag length of the domestic variables, the foreign variables and the number of the exogenous variables.}

\item{T}{: number of observations.

\if{html}{\out{<div class="sourceCode">}}\preformatted{         (m,n,p,T) are parameters which must be provided.
}\if{html}{\out{</div>}}}

\item{W}{: an (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of i-th country diag(W)=0}

\item{r_npo}{: an (m, p, n) array collecting the roots of the characteristic functions in the lag operator of the country VAR. The number of ones in each ith row of r_npo is the number of unit roots i-th country/unit.}

\item{Ao}{: an (m, m, p, n) array collecting the  coefficients of foreign variables.}

\item{Bo}{: an (m, m, p, n) array collecting the  coefficients of domestic variables.}

\item{Co}{: an (m , k+1, n) array collecting the coefficients of the deterministic components of the n countries.}

\item{Uo}{: an (T x mn) matrix of the temporally independent innovation processes}

\item{Sigmao}{: (mn x mn) matrix of the covariance matrix of the CIGVAR(m,n,p)

\if{html}{\out{<div class="sourceCode">}}\preformatted{          (W,r_npo,Ao,Bo,Uo,Sigmao) if not provided, they will be generated randomly. The default assumption is one unit root in one country. Hence m-1 cointegration relations in each country.
}\if{html}{\out{</div>}}}

\item{type}{: deterministic component "const", "none", "exog0", and "exog1" are four options}

\item{X}{: a (T x k x n) array of exogenous stationary variables.}

\item{mu}{: if type = "const" mu has the same dimension as Co. is an muV is nm vector of the means of the time series in the system}

\item{DFYflag}{: DFYflag = 0 implies foreign variables are not in the cointegration space. DFYflag = 1 allows foreign variables enter the cointegration relation.}

\item{crk}{: n vector containing the cointegration rank in each country/unit. crk is to specified for estimation of parameters.}

\item{Ncommfakt}{: number of common exogenous stochastic factors}

\item{A}{: a transformation matrix that adds the common exogenous factors to the system.}

\item{uz}{: innovations of the exogenous common factors}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a CIGVAR object containing the generated data, the parameters used and the exogenous variables.
}\if{html}{\out{</div>}}
}
\description{
This function generates data from a cointegrated global VAR process and returns an CIGVAR(m,n,p) object that is a list containing data and parameters used in the CIGVAR(m,n,p) process.
}
\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3;
res_d = CIGVARData(m=4,n=5,p=p,T=100,type="const")
max(abs(res_d$Y))
plot(ts(res_d$Y[,1:10]))
res_d$r_npo
STAT(res_d$G)
res_e = CIGVARest(res_d)
res_e$Summary
STAT(res_d$G)
plot(ts(res_d$Y[,1:10]))

}
