% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR_d.R
\name{MRCIGVARest}
\alias{MRCIGVARest}
\title{Estimation of MRCIGVAR(res)}
\usage{
MRCIGVARest(res = res_d)
}
\arguments{
\item{res}{: an MRCIVAR object of the output of MRCIGVARData}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ : an MRCIVAR object with estimated parameters and test statistics.
}\if{html}{\out{</div>}}
}
\description{
This function estimates parameters of a multi regime cointegrated global VAR(p) model. The adjustment speeds to the cointegration relations are assumed to be different in different regimes.
}
\examples{
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 3; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)

## case of n = 3, m = 2, S = 2     ## m: number of variables, n: number of countries

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),r=rep(1,3),Ncommtrend=1)
# Ao
max(abs(res_d$Y))
plot(ts(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])
max(abs(res_d$Y))
res_d$type
res_d$W
#   Ao =  res_d$Ao;  Bo =  res_d$Bo;  Co =  res_d$Co;
#   res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=300,S=2, SESVI=c(1,3,5),Ao=Ao,Bo=Bo,Co=Co)

max(res_d$Y)
res_e  = MRCIGVARest(res=res_d)
res_e$Summary

}
